requirements.txt
llama-cpp-python
pdfminer.six
python-docx
spacy
fuzzywuzzy
python-Levenshtein
PyYAML
tk  
